backend/requirements.txt
fastapi==0.123.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.38
pymysql==1.1.0  
alembic==1.12.1
pydantic==2.12.5
python-jose[cryptography]==3.3.0  
passlib[bcrypt]==1.7.4  
python-multipart==0.0.6  
stripe==7.4.0  
python-dotenv==1.0.0  
email-validator==2.0.0  